requirements.txt
python-dotenv
gspread
google-auth
PyGithub
requests
markdown